Use virtualenv instead of Docker to allow for CI integration tests

+++ bindings/pydeck/tests/dev-containers/build-and-screenshot.sh
@@ -1,15 +1,24 @@
 #!/bin/bash
 
-USAGE="$(basename "$0") [-h] [-port=] [-skip-dist-build] [-skip-jupyterlab] [-skip-jupyter-notebook] -- run current pydeck branch in JupyterLab and Jupyter Notebook
+# Ignore pip upgrade warnings
+PIP_DISABLE_PIP_VERSION_CHECK=1
+
+USAGE="$(basename "$0") [-h] [--port=] [--skip-dist-build] \
+[--skip-jupyterlab] \
+[--skip-jupyter-notebook] \
+[--skip-deckgl-build]
+  -- run current pydeck branch in JupyterLab and Jupyter Notebook
 
 where:
-    -h                      show this text
-    -port                   port on which to run local Python package index (default: 8080)
-    -skip-jupyterlab        run a JupyterLab build (default: false)
-    -skip-jupyter-notebook  run a Jupyter Notebook build (default: false)
-    -skip-dist-build        skip building a new distribution version of pydeck (default: false)"
+    -h                       show this text
+    --port                   port on which to run local Python package index (default: 8080)
+    --skip-jupyterlab        skip JupyterLab build (default: false)
+    --skip-jupyter-notebook  skip Jupyter Notebook build (default: false)
+    --skip-deckgl-build      cancels \`yarn bootstrap\`, not rebuilding deck.gl (default: false)
+    --skip-dist-build        skip building a new distribution version of pydeck (default: false)"
 
-YELLOW='\033[0;33m'
+YELLOW='\033[1;33m'
+RED='\033[1;31m'
 NC='\033[0m' # No Color
 
 for ARGUMENT in "$@"
@@ -21,51 +30,84 @@ do
   case "$KEY" in
     -h)                       echo "$USAGE" && exit ;;
     -port)                    PORT=${VALUE} ;;
-    -skip-dist-build)         SKIP_DIST_BUILD=1 ;;     
-    -skip-jupyterlab)         SKIP_JUPYTERLAB=1 ;;     
-    -skip-jupyter-notebook)   SKIP_JUPYTER_NOTEBOOK=1 ;;     
+    -skip-dist-build)         SKIP_DIST_BUILD=true ;;     
+    -skip-jupyterlab)         SKIP_JUPYTERLAB=true ;;     
+    -skip-deckgl-build)       SKIP_DECKGL_BUILD=true ;;     
+    -skip-jupyter-notebook)   SKIP_JUPYTER_NOTEBOOK=true ;;     
     *)                        echo "invalid argument $KEY" && echo "$USAGE" && exit ;;
   esac    
 done
 
 PORT=${PORT:-8080}
-SKIP_DIST_BUILD=${SKIP_DIST_BUILD:-0}
-SKIP_JUPYTERLAB=${SKIP_JUPYTERLAB:-0}
-SKIP_JUPYTER_NOTEBOOK=${SKIP_JUPYTER_NOTEBOOK:-0}
+SKIP_DIST_BUILD=${SKIP_DIST_BUILD:-false}
+SKIP_DECKGL_BUILD=${SKIP_DECKGL_BUILD:-false}
+SKIP_JUPYTERLAB=${SKIP_JUPYTERLAB:-false}
+SKIP_JUPYTER_NOTEBOOK=${SKIP_JUPYTER_NOTEBOOK:-false}
+
+function log_info {
+  printf "[`date`] $YELLOW$1$NC\n"
+}
+
+function log_error {
+  printf "[`date`] $RED$1$NC\n"
+}
+
 
 if [[ -z `pip list | grep pypiserver` ]]; then
-  printf "$YELLOW  Installing pypiserver... $NC\n"
+  log_info "Installing pypiserver..."
   pip install pypiserver
 fi
 
-# Check if port taken
-nc -z -v -G5 localhost $PORT &> /dev/null
-result=$?
-if [ "$result" == 0 ]; then
-  echo "Port $PORT occupied. Please use a different port or free this one."
-  exit 1
+if [[ -z `npm list -g | grep verdaccio` ]]; then
+  log_info "Installing verdaccio..."
+  npm install -g verdaccio
 fi
 
+function check_port {
+  nc -z -v -G5 localhost $PORT &> /dev/null
+  result=$?
+  if [ "$result" == 0 ]; then
+    log_error "Port $PORT occupied. Specify another or free this one."
+    exit 1
+  fi
+}
 
-if [ "$SKIP_DIST_BUILD" ]; then
-  printf "$YELLOW  Skipping dist build, serving builds that are present $NC\n"
+check_port
+
+if $SKIP_DECKGL_BUILD; then
+  log_info "Skipping deck.gl build"
 else
-  printf "$YELLOW  Building pydeck distribution from current branch $NC\n"
-  python ../../setup.py sdist bdist_wheel
+  log_info "Building deck.gl from current branch..." && yarn bootstrap
+fi
+
+if $SKIP_DIST_BUILD; then
+  log_info "Skipping dist build, serving builds that are present"
+else
+  pushd ../..
+  log_info "Building pydeck distribution from current branch..."
+  python setup.py sdist bdist_wheel --is-integration-build
+  popd
 fi
 
 # Starts a local pypi server
-pypiserver -p $PORT ../../dist/ &
-PYPISERVER_PID=$!
+pypi-server -p $PORT ./dist/ &
+export PYPISERVER_PID=$!
 
-function cleanup_pypi_server {
-  kill $PYPISERVER_PID
+function cleanup {
+  if [ -n "$PYPISERVER_PID" ]; then
+    kill $PYPISERVER_PID
+  fi
+  if [ -n "$JL_PID" ]; then
+    kill $JL_PID
+  fi
+  if [ -n "$JN_PID" ]; then
+    kill $JN_PID
+  fi
 }
-trap cleanup_pypi_server EXIT
+trap cleanup EXIT
 
-# # All of these variables are used within the Docker builds
-# export PYDECK_VERSION=`python -c "import pydeck; print(pydeck.__version__)"`
-# export DECKGL_VERSION=`python -c "import pydeck; print(pydeck.frontend_semver.DECKGL_SEMVER)"`
+export PYDECK_VERSION=`python -c "import pydeck; print(pydeck.__version__)"`
+export DECKGL_VERSION=`python -c "import pydeck; print(pydeck.frontend_semver.DECKGL_SEMVER)"`
 export PYPI_INSTALL_URL=http://localhost:$PORT
 
 function jupyterlab_setup {
@@ -75,10 +117,11 @@ function jupyterlab_setup {
   jupyter serverextension enable --py jupyterlab --sys-prefix
   pip install -i $PYPI_INSTALL_URL --extra-index-url https://pypi.org/simple pydeck==$PYDECK_VERSION --pre
   jupyter labextension install @jupyter-widgets/jupyterlab-manager
+  # TODO This command takes a very long time for some reason, would be great to understand why
   jupyter labextension install @deck.gl/jupyter-widget@$DECKGL_VERSION
-  jupyter lab --no-browser &> /dev/null &
-  JUPYTERLAB_PID=$!
-  return $JUPYTERLAB_PID
+  jupyter lab --no-browser &
+  PID=$!
+  return PID
 }
 
 function jupyter_notebook_setup {
@@ -87,58 +130,51 @@ function jupyter_notebook_setup {
   pip install -i $PYPI_INSTALL_URL --extra-index-url https://pypi.org/simple pydeck==$PYDECK_VERSION --pre
   jupyter nbextension install --sys-prefix --symlink --overwrite --py pydeck
   jupyter nbextension enable --sys-prefix --py pydeck
-  jupyter notebook --no-browser &> /dev/null &
-  JUPYTER_NOTEBOOK_PID=$!
-  return $JUPYTER_NOTEBOOK_PID
+  jupyter notebook --no-browser &
+  PID=$!
+  return PID
 }
 
 function activate {
+  # Use a particular Python environment
+  # We'll exit this env after the shell exits this script
   source ./env/bin/activate
-  printf "$YELLOW  Using python at `which python` $NC\n"
+  log_info "Using python at `which python`"
 }
 
 function fresh_env {
-  printf "$YELLOW  Creating new virtual envrionment $NC\n"
+  # Create a new virtualenv
+  log_info "Creating new virtual environment"
   rm -rf ./env
   python3 -m venv env
 }
 
-fresh_env()
-activate()
+fresh_env
+activate
 
-function clean_jl {
-  kill $JL_PID
-}
-trap clean_jl EXIT 
-
-function clean_jn {
-  kill $JN_PID
-}
-trap clean_jn EXIT 
-
-if [ "$SKIP_JUPYTERLAB" ]; then
-  printf "$YELLOW  Skipping JupyterLab build $NC\n"
+if $SKIP_JUPYTERLAB; then
+  log_info "Skipping JupyterLab build"
 else
-  printf "$YELLOW  Running JupyterLab build $NC\n"
-  JL_PID=`jupyterlab_setup()`
+  log_info "Running JupyterLab build"
+  log_info "Please note: This currently takes 1-2 minutes"
+  export JL_PID=`jupyterlab_setup`
 fi
 
-if [ "$SKIP_JUPYTER_NOTEBOOK" ]; then
-  printf "$YELLOW  Skipping Jupyter Notebook build $NC\n"
+if $SKIP_JUPYTER_NOTEBOOK; then
+  log_info "Skipping Jupyter Notebook build"
 else
-  printf "$YELLOW  Running Jupyter Notebook build $NC\n"
-
-  JL_PID=`jupyter_notebook_setup()`
+  log_info "Running Jupyter Notebook build"
+  export JN_PID=`jupyter_notebook_setup`
 fi
 
-control_c()
-{
+function ctrlc {
+  cleanup
   exit 0
 }
 
-trap control_c SIGINT
+trap ctrlc SIGINT
 
-printf "$YELLOW  Jupyter Notebook running at http://locahost:8888\n"
-printf "$YELLOW  JupyterLab running at http://locahost:8889\n"
-printf "$YELLOW  Ctrl+C to Exit\n"
+log_info "JupyterLab will be running at http://locahost:8888"
+log_info "Jupyter Notebook will be running at http://locahost:8889"
+log_info "Ctrl+C to Exit"
 read -r -d '' _ </dev/tty
